+++ Core/Core.csproj
@@ -55,10 +55,10 @@
     <Compile Include="src\documents\Document.cs" />
     <Compile Include="src\documents\InquirySheet.cs" />
     <Compile Include="src\documents\WarrantyCertification.cs" />
+    <Compile Include="src\ProjectManager.cs" />
     <Compile Include="src\SebeClient.cs" />
     <Compile Include="src\Core.cs" />
     <Compile Include="src\ProjectData.cs" />
-    <Compile Include="src\utils\Handler.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="src\utils\DataFile.cs" />
     <Compile Include="src\utils\Logger.cs" />
